Warning hunting.

+++ src/MEDPartitioner/MEDPARTITIONER_MeshCollectionMedXmlDriver.cxx
@@ -216,7 +216,7 @@ void MeshCollectionMedXmlDriver::write(const char* filename, ParaDomainSelector*
   //Splitting tag
   node=xmlNewChild(root_node,0,BAD_CAST "splitting",0);
   node2=xmlNewChild(node,0,BAD_CAST "subdomain",0);
-  sprintf(buff, "%d", _collection->getMesh().size());
+  sprintf(buff, "%d", (int)_collection->getMesh().size());
   xmlNewProp(node2, BAD_CAST "number", BAD_CAST buff);
   node2=xmlNewChild(node,0,BAD_CAST "global_numbering",0);
   xmlNewProp(node2, BAD_CAST "present", BAD_CAST "yes");
